Restructure test directories
