+++ lib/board.js
@@ -66,7 +66,7 @@ var Serial = {
 
   detect: function(callback) {
     var serialport = IS_TEST_MODE ?
-      require("../test/mock-serial") :
+      require("../test/util/mock-serial") :
       require("serialport");
 
     // Request a list of available ports, from
