+++ test/board-connection.js
@@ -1,5 +1,5 @@
 var sinon = require("sinon");
-var MockFirmata = require("./mock-firmata");
+var MockFirmata = require("./util/mock-firmata");
 var Board = require("../lib/board.js");
 
 var Serial = Board.__spy.Serial;
