+++ test/board.component.js
@@ -2,7 +2,7 @@ require("es6-shim");
 
 global.IS_TEST_MODE = true;
 
-var MockFirmata = require("./mock-firmata"),
+var MockFirmata = require("./util/mock-firmata"),
   five = require("../lib/johnny-five.js"),
   sinon = require("sinon"),
   Board = five.Board;
