+++ test/board.js
@@ -2,8 +2,8 @@ require("es6-shim");
 
 global.IS_TEST_MODE = true;
 
-var SerialPort = require("./mock-serial").SerialPort,
-  MockFirmata = require("./mock-firmata"),
+var SerialPort = require("./util/mock-serial").SerialPort,
+  MockFirmata = require("./util/mock-firmata"),
   five = require("../lib/johnny-five.js"),
   Repl = require("../lib/repl"),
   sinon = require("sinon"),
