+++ test/board.pins.js
@@ -1,4 +1,4 @@
-var pins = require("./mock-pins");
+var pins = require("./util/mock-pins");
 var Pins = require("../lib/board.pins.js");
 var MAX_SAFE_INTEGER = Number.MAX_SAFE_INTEGER || 9007199254740991;
