+++ test/capabilities.js
@@ -1,5 +1,5 @@
-var MockFirmata = require("./mock-firmata"),
-  pins = require("./mock-pins"),
+var MockFirmata = require("./util/mock-firmata"),
+  pins = require("./util/mock-pins"),
   five = require("../lib/johnny-five.js"),
   Board = five.Board,
   board = new Board({
