+++ test/compass.js
@@ -1,6 +1,6 @@
 var five = require("../lib/johnny-five.js"),
   sinon = require("sinon"),
-  MockFirmata = require("./mock-firmata"),
+  MockFirmata = require("./util/mock-firmata"),
   Board = five.Board,
   Compass = five.Compass;
 
@@ -114,4 +114,3 @@ exports["Invalid or missing controller"] = {
     test.done();
   },
 };
-
