+++ test/distance.js
@@ -1,4 +1,4 @@
-var MockFirmata = require("./mock-firmata"),
+var MockFirmata = require("./util/mock-firmata"),
   five = require("../lib/johnny-five.js"),
   events = require("events"),
   sinon = require("sinon"),
