+++ test/pin.js
@@ -1,7 +1,7 @@
 global.IS_TEST_MODE = true;
 
 var five = require("../lib/johnny-five.js"),
-  MockFirmata = require("./mock-firmata"),
+  MockFirmata = require("./util/mock-firmata"),
   sinon = require("sinon"),
   events = require("events"),
   Board = five.Board,
