+++ test/ping.js
@@ -1,6 +1,6 @@
-var MockFirmata = require("./mock-firmata"),
+var MockFirmata = require("./util/mock-firmata"),
   sinon = require("sinon"),
-  pins = require("./mock-pins"),
+  pins = require("./util/mock-pins"),
   five = require("../lib/johnny-five.js"),
   Board = five.Board,
   Ping = five.Ping,
