+++ test/util/mock-firmata.js
