+++ test/util/mock-serial.js
